backend/requirements.txt
flask
flask-cors
requests
pillow
psutil
sounddevice
numpy
openai-whisper
scipy
pyttsx3
textblob
google-api-python-client
google-auth-httplib2
google-auth-oauthlib